% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ym.R
\name{combine_ym}
\alias{combine_ym}
\title{combine_ym}
\usage{
combine_ym(dataset, column)
}
\arguments{
\item{dataset}{data set you want to add a column to.}

\item{column}{this can be a column of the data set using the $ operator or a vector that will be
added to dataset as a cloumn.}
}
\value{
dataset with the colums: year_created, month_created.
}
\description{
fixes dates and months from "Created" or "created_year_month" columns to allow for easy query.
}
