% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_month.R
\name{convert_month}
\alias{convert_month}
\title{convert_month}
\usage{
convert_month(x)
}
\arguments{
\item{x}{this is a vector, matrix, or subsetted matrix/dataframe in the form of a vector}
}
\value{
a vector of full name months
}
\description{
Converts any numbers 1:12 to months of the year character strings
}
