% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_year.R
\name{fix_year}
\alias{fix_year}
\title{fix_year}
\usage{
fix_year(v)
}
\arguments{
\item{v}{this is a vector using the $ from a data frame or subsetted from a matrix or data frame}
}
\value{
a vector of dates as character strings
}
\description{
takes the character string column of the Miami311 data csv file, "Created" or gathered API data, "created_year_month", and filters out the year part of the character string.
}
