% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellofun.R
\name{geteR}
\alias{geteR}
\title{geteR}
\usage{
geteR(name = "Insert Your Name", by = "day")
}
\arguments{
\item{name}{Insert your name in quotation marks. It is required to start the function.}

\item{by=}{If (by="day"), this means your data is gathered once a day ONLY.
If (by = "time"), this means your data is gathered multiple times in one day
The default is by = "day"
If running more than once in a day due to a mistake, set by = "time". This will stop an error
message from appearing.}
}
\value{
Miami311API: object in global environment

miami311API.csv: data file with the current date's API data

mapi_db: this is the name of the database in R created from the function
}
\description{
Gathers Miami json API data from the Miami-Dade County website and stores it in to RSQLite database.
}
